feat(brewfiles/base/Brewfile): add GitHub Actions and Bun extensions to VSCode setup
This commit introduces two new Visual Studio Code extensions to the base
Brewfile configuration. The `vscode-github-actions` extension is added to
facilitate the development and management of GitHub Actions workflows
directly within the editor. Additionally, the `bun-vscode` extension is
included to provide support for the Bun JavaScript runtime, which is
gaining popularity in the development community. These additions aim to
enhance the development experience by integrating useful tools and
features into the editor setup.